+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/requests/ListServiceConfigsParams.java
@@ -1,0 +1,27 @@
+package cloud.pangeacyber.pangea.ai_guard.requests;
+
+import cloud.pangeacyber.pangea.BaseRequest;
+import cloud.pangeacyber.pangea.ai_guard.models.ServiceConfigListFilter;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
+import lombok.Value;
+import lombok.extern.jackson.Jacksonized;
+
+@Builder
+@Jacksonized
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@Value
+public final class ListServiceConfigsParams extends BaseRequest {
+
+	ServiceConfigListFilter filter;
+	String last;
+	String order;
+	String orderBy;
+	Integer size;
+}